+++ src/routes/__tests__/graphql.test.js
@@ -1,0 +1,11 @@
+const graphql = require('../graphql');
+
+jest.mock('graphql-server-micro', () => ({
+  microGraphql: d => d,
+}));
+
+jest.mock('../../graphql/schema', () => ({ schema: 'mock' }));
+
+it('should return graphql arguments', () => {
+  expect(graphql()).toMatchSnapshot();
+});