+++ src/routes/home.js
@@ -10,6 +10,10 @@ const home = (req, res) =>
 
     Endpoint:
     [GET/POST] /graphql
+
+
+
+    https://github.com/evenchange4/micro-medium-api
 `,
   );